fixes hidden errors in non awaits for chai expects
