+++ package.json
@@ -1,52 +1,46 @@
 {
-  "name": "hardhat-project",
-  "dependencies": {
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.8.0",
-    "@mui/material": "^5.8.0",
-    "@openzeppelin/contracts": "^4.6.0",
-    "clsx": "^1.1.1",
-    "date-fns": "^2.28.0",
-    "next": "^12.1.6",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "react-router-dom": "^6.3.0",
-    "sass": "^1.52.1"
-  },
-  "devDependencies": {
-    "@nomiclabs/hardhat-ethers": "^2.0.6",
-    "@nomiclabs/hardhat-waffle": "^2.0.3",
-    "@typechain/ethers-v5": "^10.0.0",
-    "@typechain/hardhat": "^6.0.0",
-    "@types/chai": "^4.3.1",
-    "@types/mocha": "^9.1.1",
-    "@types/node": "^17.0.35",
-    "@types/react": "^18.0.9",
-    "chai": "^4.3.6",
-    "eslint": "^8.16.0",
-    "eslint-config-standard": "^17.0.0",
-    "ethereum-waffle": "^3.4.4",
-    "ethers": "^5.6.7",
-    "hardhat": "^2.9.5",
-    "prettier": "^2.6.2",
-    "ts-node": "^10.7.0",
-    "typechain": "^8.0.0",
-    "typescript": "^4.6.4"
-  },
-  "scripts": {
-    "clean": "npx hardhat clean",
-    "compile": "npx hardhat compile",
-    "test:publishOptionRequest": "npx hardhat test test/publishOptionRequest.ts",
-    "test:withdrawOptionRequest": "npx hardhat test test/withdrawOptionRequest.ts",
-    "test:createOption": "npx hardhat test test/createOption.ts",
-    "test:exerciseOption": "npx hardhat test test/exerciseOption.ts",
-    "test:cancelOption": "npx hardhat test test/cancelOption.ts",
-    "test:all": "npx hardhat test",
-    "node": "npx hardhat node",
-    "deploy": "npx hardhat run scripts/deploy.ts --network localhost",
-    "next:dev": "next dev -p 3003",
-    "next:build": "next build",
-    "next:start": "next start"
-  }
+    "name": "hardhat-project",
+    "dependencies": {
+        "@emotion/react": "^11.9.0",
+        "@emotion/styled": "^11.8.1",
+        "@mui/icons-material": "^5.8.0",
+        "@mui/material": "^5.8.0",
+        "@openzeppelin/contracts": "^4.6.0",
+        "clsx": "^1.1.1",
+        "next": "^12.1.6",
+        "react": "^18.1.0",
+        "react-dom": "^18.1.0",
+        "react-router-dom": "^6.3.0",
+        "sass": "^1.52.1"
+    },
+    "devDependencies": {
+        "@nomiclabs/hardhat-ethers": "^2.0.6",
+        "@nomiclabs/hardhat-waffle": "^2.0.3",
+        "@typechain/ethers-v5": "^10.0.0",
+        "@typechain/hardhat": "^6.0.0",
+        "@types/chai": "^4.3.1",
+        "@types/mocha": "^9.1.1",
+        "@types/node": "^17.0.35",
+        "@types/react": "^18.0.9",
+        "chai": "^4.3.6",
+        "eslint": "^8.16.0",
+        "eslint-config-standard": "^17.0.0",
+        "ethereum-waffle": "^3.4.4",
+        "ethers": "^5.6.7",
+        "hardhat": "^2.9.5",
+        "prettier": "^2.6.2",
+        "ts-node": "^10.7.0",
+        "typechain": "^8.0.0",
+        "typescript": "^4.6.4"
+    },
+    "scripts": {
+        "clean": "npx hardhat clean",
+        "compile": "npx hardhat compile",
+        "test": "npx hardhat test",
+        "node": "npx hardhat node",
+        "deploy": "npx hardhat run scripts/deploy.ts --network localhost",
+        "next:dev": "next dev",
+        "next:build": "next build",
+        "next:start": "next start"
+    }
 }
